4_Hands_On_Code/pyproject.toml
[build-system]
requires = ["setuptools>=82.0.1"]
build-backend = "setuptools.build_meta"

[project]
name = "common"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["common*"]